+++ src/base/SiteGroupProcessor.php
@@ -1,0 +1,47 @@
+<?php
+/**
+ * Architect plugin for Craft CMS 3.x
+ *
+ * CraftCMS plugin to generate content models from JSON data.
+ *
+ * @link      https://pennebaker.com
+ * @copyright Copyright (c) 2018 Pennebaker
+ */
+
+namespace pennebaker\architect\base;
+
+use Craft;
+use craft\models\SiteGroup;
+
+/**
+ * SiteGroupProcessor defines the common interface to be implemented by plugin classes.
+ *
+ * @author    Pennebaker
+ * @package   Architect
+ * @since     2.0.0
+ */
+class SiteGroupProcessor extends Processor
+{
+    /**
+     * @param array $item
+     *
+     * @return array
+     */
+    public function parse(array $item)
+    {
+        return [new SiteGroup($item), null];
+    }
+
+    /**
+     * @param mixed $item
+     * @param bool $update
+     *
+     * @return bool|object
+     *
+     * @throws \Throwable
+     */
+    public function save($item, bool $update = false)
+    {
+        return Craft::$app->sites->saveGroup($item);
+    }
+}